Restructure wallet
